Add elements module and formatted output
Adds a new elements module in data that contains all of the known
elements. Adds new logic to the ability formatter (which is responsible
for the actual bot output for boon commands) to include element info
when it is available. Still works where an element is not defined,
whether because the data is not yet updated or because the boon has no
element (as in the case of infusions). Updates element data in
Aphrodites boons so that this is actually testable end-to-end.