Implements prereq command
Adds an extension of the boon command which responds to command syntax
like !flutter flourish prereqs. Code satisfies new scenarios which test
prereqs queries. Adds a new formatter to construct prereq strings. As
prereq definitions ask almost every god to depend on almost every other
god, it seemed important in testing for the formatter to not be called
until runtime / in the handler. Attempts to add the strings to god
definitions choked on incomplete data in the prereq arrays.