Fix syntax bug in boon requirements command
Previously, the boon command supported both user requests for boon info
in the form '!boon name' and '!boonName' with any capitalization.
However, the deeper boon requirements command was failing to recognize
'!boonName reqs', responding instead only to '!boon name reqs' with
spacing.

The bug, it turns out, is that we construct the boon name matcher twice;
once for boon info in the abilityExpressionMap, and once again for boon
reqs info in abilityPrereqMap. In the ExpressionMap, we converted
whitespace to ' *' 0-or-more matcher, but failed to do so in the
PrereqMap.

This commit fixes this issue. However, this betrays an unnecessary
fragility of the code. It would be better to construct these matching
strings one time and then to share them across both structures.