Implements Chaos command
Adds bot capability to handle messages in the form !curse blessing and
responds with information about a Chaos combo. Tests this new
functionality in a new chaos feature. Also adds support and testing for
querying Chaos' infusion, Chant, by various routes (!infusions cosmic,
!chaos infusion, !chant).

In particular, this change alters the data model of both ChaosBlessing
and ChaosCurse. Almost all blessings and curses have a range of possible
values, and all curse durations are also ranges. Additionally, some
values should be expressed as percentages while others are simple
numbers. Finally, blessings have rarities (like Boons) that are
associated with some multiplier to apply to the value ranges.

With this change, all string values are removed in favor of numeric
ranges. In almost all cases, strings were used to represent percentages;
this is a manipulation we can do with formatting logic at handler time.
In the few other cases, the string was a constant "value" across all
rarities (such as a boon rarity) and could just be hardcoded into info.

While all new functionality is confirmed to work through testing, there
are still several steps to clean this code up. At present, all
formatting is being done within the command code. Some or all should be
decomposable to formatter helper functions. Additionally, there is a
fair bit of redundancy between curses and blessings that could likely be
reduced.

In the long run, it's possible some of what we are doing with blessings
and curses could be extended to boons, as many of them follow similar
logic for rarity scaling and percentage formatting.